requirements.txt
certifi (2017.7.27.1)
chardet (3.0.4)
coreapi (2.3.3)
coreschema (0.0.4)
Django (1.11.6)
django-crispy-forms (1.6.1)
django-filter (1.0.4)
django-formtools (2.1)
django-guardian (1.4.9)
django-reversion (2.0.10)
djangorestframework (3.7.0)
future (0.16.0)
httplib2 (0.10.3)
idna (2.6)
itypes (1.1.0)
Jinja2 (2.9.6)
Markdown (2.6.9)
MarkupSafe (1.0)
olefile (0.44)
Pillow (4.3.0)
pip (9.0.1)
pytz (2017.2)
requests (2.18.4)
setuptools (28.8.0)
six (1.11.0)
uritemplate (3.0.0)
urllib3 (1.22)
XlsxWriter (1.0.0)
xlwt (1.3.0)
